refactor project using generics (api change)
