+++ src/lib.rs
@@ -42,9 +42,7 @@ pub mod api;
 pub mod iter;
 pub mod parser;
 
-use crate::api::{BlockIterator, FBlock, FConnectedBlock, SBlock, SConnectedBlock};
-use bitcoin::hashes::hex::{FromHex, ToHex};
-use bitcoin::Txid;
+use crate::api::{BlockIterator, FBlock, FConnectedBlock, FTransaction, FromHex, SBlock, SConnectedBlock, STransaction, ToHex, Txid, FConnectedTransaction, SConnectedTransaction};
 use pyo3::prelude::*;
 use pyo3::PyIterProtocol;
 use pyo3::Python;
@@ -69,7 +67,7 @@ impl BitcoinDB {
 
     #[pyo3(text_signature = "($self, height, /)")]
     fn get_block_full(&self, height: i32, py: Python) -> PyResult<PyObject> {
-        match self.db.get_block_full(height) {
+        match self.db.get_block::<FBlock>(height) {
             Ok(block) => Ok(pythonize(py, &block)?),
             Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
         }
@@ -77,7 +75,7 @@ impl BitcoinDB {
 
     #[pyo3(text_signature = "($self, height, /)")]
     fn get_block_simple(&self, height: i32, py: Python) -> PyResult<PyObject> {
-        match self.db.get_block_simple(height) {
+        match self.db.get_block::<SBlock>(height) {
             Ok(block) => Ok(pythonize(py, &block)?),
             Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
         }
@@ -85,7 +83,7 @@ impl BitcoinDB {
 
     #[pyo3(text_signature = "($self, height, /)")]
     fn get_block_full_connected(&self, height: i32, py: Python) -> PyResult<PyObject> {
-        match self.db.get_block_full_connected(height) {
+        match self.db.get_block_connected::<FConnectedBlock>(height) {
             Ok(block) => Ok(pythonize(py, &block)?),
             Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
         }
@@ -93,7 +91,7 @@ impl BitcoinDB {
 
     #[pyo3(text_signature = "($self, height, /)")]
     fn get_block_simple_connected(&self, height: i32, py: Python) -> PyResult<PyObject> {
-        match self.db.get_block_simple_connected(height) {
+        match self.db.get_block_connected::<SConnectedBlock>(height) {
             Ok(block) => Ok(pythonize(py, &block)?),
             Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
         }
@@ -101,7 +99,7 @@ impl BitcoinDB {
 
     #[pyo3(text_signature = "($self, height, /)")]
     fn get_block_header(&self, height: usize, py: Python) -> PyResult<PyObject> {
-        match self.db.get_block_header(height) {
+        match self.db.get_header(height) {
             Ok(block) => Ok(pythonize(py, &block)?),
             Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
         }
@@ -126,7 +124,7 @@ impl BitcoinDB {
     #[pyo3(text_signature = "($self, txid, /)")]
     fn get_height_from_txid(&self, txid: String) -> PyResult<i32> {
         if let Ok(txid) = Txid::from_hex(&txid) {
-            match self.db.get_block_height_of_transaction(&txid) {
+            match self.db.get_height_of_transaction(&txid) {
                 Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
                 Ok(h) => Ok(h),
             }
@@ -140,7 +138,7 @@ impl BitcoinDB {
     #[pyo3(text_signature = "($self, txid, /)")]
     fn get_transaction_full(&self, txid: String, py: Python) -> PyResult<PyObject> {
         if let Ok(txid) = Txid::from_hex(&txid) {
-            match self.db.get_transaction_full(&txid) {
+            match self.db.get_transaction::<FTransaction>(&txid) {
                 Ok(t) => Ok(pythonize(py, &t)?),
                 Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
             }
@@ -154,7 +152,7 @@ impl BitcoinDB {
     #[pyo3(text_signature = "($self, txid, /)")]
     fn get_transaction_simple(&self, txid: String, py: Python) -> PyResult<PyObject> {
         if let Ok(txid) = Txid::from_hex(&txid) {
-            match self.db.get_transaction_simple(&txid) {
+            match self.db.get_transaction::<STransaction>(&txid) {
                 Ok(t) => Ok(pythonize(py, &t)?),
                 Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
             }
@@ -168,7 +166,7 @@ impl BitcoinDB {
     #[pyo3(text_signature = "($self, txid, /)")]
     fn get_transaction_full_connected(&self, txid: String, py: Python) -> PyResult<PyObject> {
         if let Ok(txid) = Txid::from_hex(&txid) {
-            match self.db.get_transaction_full_connected(&txid) {
+            match self.db.get_transaction_connected::<FConnectedTransaction>(&txid) {
                 Ok(t) => Ok(pythonize(py, &t)?),
                 Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
             }
@@ -182,7 +180,7 @@ impl BitcoinDB {
     #[pyo3(text_signature = "($self, txid, /)")]
     fn get_transaction_simple_connected(&self, txid: String, py: Python) -> PyResult<PyObject> {
         if let Ok(txid) = Txid::from_hex(&txid) {
-            match self.db.get_transaction_simple_connected(&txid) {
+            match self.db.get_transaction_connected::<SConnectedTransaction>(&txid) {
                 Ok(t) => Ok(pythonize(py, &t)?),
                 Err(e) => Err(pyo3::exceptions::PyException::new_err(e.to_string())),
             }
@@ -249,7 +247,7 @@ struct SBlockIteratorArray {
 impl SBlockIteratorArray {
     fn new(db: &api::BitcoinDB, heights: Vec<u32>) -> Self {
         SBlockIteratorArray {
-            iter: db.get_block_simple_iter_arr(heights),
+            iter: db.iter_heights::<SBlock>(heights),
         }
     }
 }
@@ -284,7 +282,7 @@ struct FBlockIteratorArray {
 impl FBlockIteratorArray {
     fn new(db: &api::BitcoinDB, heights: Vec<u32>) -> Self {
         FBlockIteratorArray {
-            iter: db.get_block_full_iter_arr(heights),
+            iter: db.iter_heights::<FBlock>(heights),
         }
     }
 }
@@ -319,7 +317,7 @@ struct FBlockIteratorSequential {
 impl FBlockIteratorSequential {
     fn new(db: &api::BitcoinDB, start: u32, end: u32) -> Self {
         FBlockIteratorSequential {
-            iter: db.get_block_full_iter_seq(start, end),
+            iter: db.iter_block::<FBlock>(start, end),
         }
     }
 }
@@ -354,7 +352,7 @@ struct SBlockIteratorSequential {
 impl SBlockIteratorSequential {
     fn new(db: &api::BitcoinDB, start: u32, end: u32) -> SBlockIteratorSequential {
         SBlockIteratorSequential {
-            iter: db.get_block_simple_iter_seq(start, end),
+            iter: db.iter_block::<SBlock>(start, end),
         }
     }
 }
@@ -389,7 +387,7 @@ struct FConnectedBlockIterator {
 impl FConnectedBlockIterator {
     fn new(db: &api::BitcoinDB, end: u32) -> FConnectedBlockIterator {
         FConnectedBlockIterator {
-            iter: db.get_block_full_connected_iter(end),
+            iter: db.iter_block_full_connected(end),
         }
     }
 }
@@ -424,7 +422,7 @@ struct SConnectedBlockIterator {
 impl SConnectedBlockIterator {
     fn new(db: &api::BitcoinDB, end: u32) -> SConnectedBlockIterator {
         SConnectedBlockIterator {
-            iter: db.get_block_simple_connected_iter(end),
+            iter: db.iter_block_simple_connected(end),
         }
     }
 }
